Updated `postinst` script to symlink to appbundle created binstubs rather than rubygems binstubs in embedded directory.
The rubygems binstubs do not account for GEM_HOME or GEM_PATHS existing
in the invoking shell. This creates problems on systems with multiple
ruby versions. If GEM_HOME or GEM_PATH is set the rubygems embedded
binstubs will attempt to use this external gem source which will
generally result in errors for the user.

Signed-off-by: Miah Johnson <miah@chia-pet.org>